refactor(app): перенести код конфигурации приложения в папку app
Jira issue: VEGA-865
