+++ package.json
@@ -73,6 +73,7 @@
     "commitizen": "^4.2.1",
     "concurrently": "^5.1.0",
     "cross-env": "^7.0.2",
+    "cross-fetch": "^3.0.6",
     "cz-customizable": "^6.3.0",
     "dotenv": "^8.2.0",
     "eslint-config-ts-important-stuff": "^1.0.0",
